revision: link subcommand
